# Avoid using public_send on Context#[]