@rmosolgo 2023-09-13
Hi! Thanks for sharing what you found in such a detailed write-up. I've been working hard to find little things to improve in query execution and I'm glad someone _else_ noticed 😆 

I agree that using something other than `public_send` is good here. I guessed that calls to runtime state methods on `query.context...` were pretty rare. Instead, I thought that _writes_ were more common, since the runtime _always_ writes to those values, not knowing if they would be used or not. 

I suspect that adding a `Hash` will make writes slower, since the incoming key must be hashed before the value can be written. One way to reduce that overhead would be to use `compare_by_identity`, but another approach might be to _not_ used a Hash, but instead, use hardcoded method calls instead of public send. 

I tried a patch like this: 

```
class Query
    class Context
      prepend(Module.new do
        def [](key)
          if @scoped_context.key?(key)
            @scoped_context[key]
          elsif @provided_values.key?(key)
            @provided_values[key]
          elsif key == :current_path
            current_path
          elsif key == :current_argument
            query_runtime_state.current_argument
          elsif key == :current_field
            query_runtime_state.current_field
          elsif key == :current_object
            query_runtime_state.current_object
          else
            # not found
            nil
          end
        end

        def query_runtime_state
          (current_runtime_state = Thread.current[:__graphql_runtime_info]) &&
            (query_runtime_state = current_runtime_state[@query])
        end
      end)
    end
  end
```

It removes the `Set#include?` call and the `.public_send`. 

Then, I ran the benchmarks three times: once on v2.1.0 from rubygems, once with the Hash-based patch above, and once with the `elsif ...`-based patch. Here were my local results: 

```
GraphQL v2.1.0
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      0.911  (± 0.0%) i/s -     28.000  in  30.773497s

GraphQL v2.1.0-with-hash-storage
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      0.837  (± 0.0%) i/s -     26.000  in  31.073932s
       
GraphQL v2.1.0-with-hardcoded-calls
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      0.987  (± 0.0%) i/s -     30.000  in  30.421189s
```

So they're all very close, but the hard-coded calls approach was even a _bit_ faster. 

What do you think of an approach like that? 

If we choose it, a couple other places would need updates: 

- https://github.com/rmosolgo/graphql-ruby/blob/2781f4fbe3c6d9e0a1399534aeaf17e070047778/lib/graphql/query/context.rb#L168
- https://github.com/rmosolgo/graphql-ruby/blob/master/lib/graphql/query/context.rb#L186

Then `RUNTIME_METADATA_KEYS` could be removed entirely. What do you think?