@swalkinshaw 2023-09-13
Are these benchmarks with YJIT? Ideally we always compare non YJIT with YJIT as well because there can be different characteristics