@rmosolgo 2023-09-13
Good call, here are my local results with YJIT enabled: 

```
GraphQL v2.1.0
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      2.778  (± 0.0%) i/s -     84.000  in  30.259305s

GraphQL v2.1.0-with-hash-storage
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      2.405  (± 0.0%) i/s -     73.000  in  30.374050s

GraphQL v2.1.0-with-hardcoded-calls
Warming up --------------------------------------
       run benchmark     1.000  i/100ms
Calculating -------------------------------------
       run benchmark      2.924  (± 0.0%) i/s -     88.000  in  30.130306s
```

(Also -- the CI failures are fixed on master, sorry about that!)