@rmosolgo 2023-09-13
I tried out that approach here: https://github.com/rmosolgo/graphql-ruby/pull/4634