@mtsmfm 2023-09-15
> I suspect that adding a Hash will make writes slower, since the incoming key must be hashed before the value can be written. 

Good point. I agree.

> What do you think of an approach like that?

Unfortunately the code doesn't look cool though, if it's the fastest way I don't agree with it 👍 

Let me close this PR.